Add missing annotation @Configuration
Hi, I found that there may be some minor improvements about annotations in your code.

It would be more appropriate to add @Configuration annotation at the class level. The @repository annotation indicates that the class provides the mechanism for configuration, which can be handed over to the container class for management.

For better understanding and maintainability of a large application, adding @Configuration is a better choice.

Hope my suggestions are useful!